+++ MetX/MetX.Tests/Standard/Generation/CSharp/Project/ModifierTests_CompilerGeneratedFilesOutputPath.cs
@@ -1,33 +1,33 @@
 ﻿using System.IO;
-using MetX.Standard.Generation.CSharp.Project;
+using MetX.Tests.Standard.Generation.CSharp.Project.Pieces;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace MetX.Tests.Standard
+namespace MetX.Tests.Standard.Generation.CSharp.Project
 {
     [TestClass]
     // ReSharper disable once InconsistentNaming
-    public class CSharpProjectModifierTests_CompilerGeneratedFilesOutputPath
+    public class ModifierTests_CompilerGeneratedFilesOutputPath
     {
-        public const string PiecesDirectory = @"Standard\ProjectPieces\GenerateToPath";
+        public const string Area = @"GenerateToPath";
         
         [TestMethod]
         public void CompilerGeneratedFilesOutputPath_GenerateToPathMissing()
         {
-            Assert.IsTrue(GetProjectPiece("Missing").PropertyGroups.CompilerGeneratedFilesOutputPathMissing);
-            Assert.IsNull(GetProjectPiece("Missing").PropertyGroups.CompilerGeneratedFilesOutputPath);
+            Assert.IsTrue(Piece.Get("Missing", Area).PropertyGroups.CompilerGeneratedFilesOutputPathMissing);
+            Assert.IsNull(Piece.Get("Missing", Area).PropertyGroups.CompilerGeneratedFilesOutputPath);
         }
         
         [TestMethod]
         public void CompilerGeneratedFilesOutputPath_PropertyGroupMissing()
         {
-            Assert.IsTrue(GetProjectPiece("Missing").PropertyGroups.CompilerGeneratedFilesOutputPathMissing);
-            Assert.IsNull(GetProjectPiece("Missing").PropertyGroups.CompilerGeneratedFilesOutputPath);
+            Assert.IsTrue(Piece.Get("Missing", Area).PropertyGroups.CompilerGeneratedFilesOutputPathMissing);
+            Assert.IsNull(Piece.Get("Missing", Area).PropertyGroups.CompilerGeneratedFilesOutputPath);
         }
 
         [TestMethod]
         public void CompilerGeneratedFilesOutputPath_EqualsGenerated()
         {
-            var project = GetProjectPiece("EqualsGenerated");
+            var project = Piece.Get("EqualsGenerated", Area);
             Assert.IsFalse(project.PropertyGroups.CompilerGeneratedFilesOutputPathMissing);
             Assert.AreEqual("Generated", project.PropertyGroups.CompilerGeneratedFilesOutputPath);
         }
@@ -35,7 +35,7 @@ namespace MetX.Tests.Standard
         [TestMethod]
         public void CompilerGeneratedFilesOutputPath_Blank()
         {
-            var project = GetProjectPiece("EqualsBlank");
+            var project = Piece.Get("EqualsBlank", Area);
             Assert.IsFalse(project.PropertyGroups.CompilerGeneratedFilesOutputPathMissing);
             Assert.AreEqual(null, project.PropertyGroups.CompilerGeneratedFilesOutputPath);
         }
@@ -43,7 +43,7 @@ namespace MetX.Tests.Standard
         [TestMethod]
         public void SetCompilerGeneratedFilesOutputPath_GeneratedToBlank() // Blank or null means remove that node
         {
-            var project = GetProjectPiece("EqualsGenerated");
+            var project = Piece.Get("EqualsGenerated", Area);
             Assert.IsFalse(project.PropertyGroups.CompilerGeneratedFilesOutputPathMissing);
             project.PropertyGroups.CompilerGeneratedFilesOutputPath = "";
             Assert.AreEqual(null, project.PropertyGroups.CompilerGeneratedFilesOutputPath);
@@ -53,7 +53,7 @@ namespace MetX.Tests.Standard
         [TestMethod]
         public void SetCompilerGeneratedFilesOutputPath_GeneratedToNullAsBlank()
         {
-            var project = GetProjectPiece("EqualsGenerated");
+            var project = Piece.Get("EqualsGenerated", Area);
             Assert.IsFalse(project.PropertyGroups.CompilerGeneratedFilesOutputPathMissing);
             project.PropertyGroups.CompilerGeneratedFilesOutputPath = null;
             Assert.AreEqual(null, project.PropertyGroups.CompilerGeneratedFilesOutputPath);
@@ -63,7 +63,7 @@ namespace MetX.Tests.Standard
         [TestMethod]
         public void SetCompilerGeneratedFilesOutputPath_MissingToGenerated()
         {
-            var project = GetProjectPiece("Missing");
+            var project = Piece.Get("Missing", Area);
             Assert.IsTrue(project.PropertyGroups.CompilerGeneratedFilesOutputPathMissing);
             var generated = "Generated";
             project.PropertyGroups.CompilerGeneratedFilesOutputPath = generated;
@@ -74,19 +74,11 @@ namespace MetX.Tests.Standard
         [TestMethod]
         public void SetGenerateToPathWhen_MissingToNull() // Don't add it if there's no value
         {
-            var project = GetProjectPiece("Missing");
+            var project = Piece.Get("Missing", Area);
             Assert.IsTrue(project.PropertyGroups.CompilerGeneratedFilesOutputPathMissing);
             project.PropertyGroups.CompilerGeneratedFilesOutputPath = "";
             Assert.AreEqual(null, project.PropertyGroups.CompilerGeneratedFilesOutputPath);
             Assert.IsFalse(project.PropertyGroups.CompilerGeneratedFilesOutputPathMissing);
         }
-
-        private static Modifier GetProjectPiece(string pieceName)
-        {
-            var filePath = $@"{PiecesDirectory}\{pieceName}.xml";
-            Assert.IsTrue(File.Exists(filePath));
-            return Modifier.LoadFile(filePath);
-        }
-
     }
 }
