+++ MetX/MetX.Tests/Standard/Generation/CSharp/Project/ModifierTests_Target_AddSource.cs
@@ -1,69 +1,41 @@
-﻿using System.Formats.Asn1;
-using System.IO;
+﻿using System.IO;
 using System.Xml;
 using MetX.Standard.Generation.CSharp.Project;
+using MetX.Tests.Standard.Generation.CSharp.Project.Pieces;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace MetX.Tests.Standard
+namespace MetX.Tests.Standard.Generation.CSharp.Project
 {
     [TestClass]
-    public class Project_PackageReference
+    public class ModifierTests_Target_AddSource
     {
         public const string EmptyClientXmlFilePath = @"Standard\ProjectPieces\EmptyClient.xml";
         public const string FullClientXmlFilePath = @"Standard\ProjectPieces\FullClient.xml";
         
         [TestMethod]
-        public void InsertOne_NotPresent()
+        public void Target_FromScratch_RemovesWhenPresent()
         {
-            var modifier = GetEmptyClient();
-            var packageName = "Microsoft.CodeAnalysis.Common";
-            var version = "3.9.0";
-            var packageReference = modifier.ItemGroup.PackageReference.GetOrInsert(packageName, version);
+            var modifier = Piece.Get("Full", null);
+            modifier.Targets.Insert();
+            modifier.Targets.Remove(); 
             
-            Assert.IsNotNull(packageReference);
-            Assert.AreEqual(packageName, packageReference.Attributes["Include"]?.Value);
-            Assert.IsNotNull(version, packageReference.Attributes["Version"]?.Value);
-        }
-        
-        [TestMethod]
-        public void InsertOne_NotPresentWithAssets()
-        {
-            var modifier = GetEmptyClient();
-            var packageName = "Microsoft.CodeAnalysis.Analyzers";
-            var version = "3.3.2";
-            var privateAssets = "all";
-            var includeAssets = "runtime; build; native; contentfiles; analyzers; buildtransitive";
-            var packageReference = modifier.ItemGroup.PackageReference.GetOrInsert(packageName, version, privateAssets, includeAssets);
+            Assert.IsNull(modifier.Targets.AddSourceGeneratedFiles);
+            Assert.IsNull(modifier.Targets.RemoveSourceGeneratedFiles);
             
-            Assert.IsNotNull(packageReference);
-            Assert.AreEqual(packageName, packageReference.Attributes["Include"]?.Value);
-            Assert.AreEqual(version, packageReference.Attributes["Version"]?.Value);
-            
-            Assert.AreEqual(2, packageReference.ChildNodes.Count);
-            Assert.AreEqual("PrivateAssets", packageReference.ChildNodes[0]?.Name);
-            Assert.AreEqual(privateAssets, packageReference.ChildNodes[0]?.InnerText);
-            
-            Assert.AreEqual("IncludeAssets", packageReference.ChildNodes[1]?.Name);
-            Assert.AreEqual(includeAssets, packageReference.ChildNodes[1]?.InnerText);
-        }
-        
-        [TestMethod]
-        public void InsertOne_Present()
-        {
-            var modifier = GetFullClient();
-            var packageName = "Microsoft.CodeAnalysis.Common";
-            var version = "3.9.0";
-            var packageReference = modifier.ItemGroup.PackageReference.GetOrInsert(packageName, version);
-            
-            Assert.IsNotNull(packageReference);
-            Assert.AreEqual(packageName, packageReference.Attributes["Include"]?.Value);
-            Assert.IsNotNull(version, packageReference.Attributes["Version"]?.Value);
+            var innerXml = modifier.Document.InnerXml;
+            var innerXmlFormatted = innerXml.Replace("><", ">\n\t<");
+            Assert.IsFalse(innerXml.Contains(@"Generated\**"), innerXmlFormatted);
+            Assert.IsFalse(innerXml.Contains(@"AddSourceGeneratedFiles"), innerXmlFormatted);
+            Assert.IsFalse(innerXml.Contains(@"RemoveSourceGeneratedFiles"), innerXmlFormatted);
+            Assert.IsNotNull(modifier.Document.SelectSingleNode(XPaths.TargetByName("SomePreExistingTarget1")));
+            Assert.IsNotNull(modifier.Document.SelectSingleNode(XPaths.TargetByName("SomePreExistingTarget2")));
+            Assert.IsNotNull(modifier.Document.SelectSingleNode(XPaths.TargetByName("SomePreExistingTarget3")));
         }
         
         [TestMethod]
         public void Target_FromScratch_RemovesWhenNotPresent()
         {
-            var modifier = GetEmptyClient();
+            var modifier = Piece.Get("Empty", null);
             modifier.Targets.Remove(); 
             
             Assert.IsNull(modifier.Targets.AddSourceGeneratedFiles);
@@ -80,7 +52,7 @@ namespace MetX.Tests.Standard
         [TestMethod]
         public void Target_FromScratch_InsertsWhenNotPresent()
         {
-            var modifier = GetEmptyClient();
+            var modifier = Piece.Get("Empty", null);
             modifier.Targets.Insert(); 
             
             Assert.IsNotNull(modifier.Targets);
@@ -98,7 +70,7 @@ namespace MetX.Tests.Standard
         [TestMethod]
         public void Target_FromScratch_GeneratesCorrectInnerXml()
         {
-            var modifier = GetEmptyClient();
+            var modifier = Piece.Get("Empty", null);
             modifier.Targets.Insert();
             Assert.IsNotNull(modifier.Targets.AddSourceGeneratedFiles);
             Assert.IsNotNull(modifier.Targets.AddSourceGeneratedFiles.TargetElement);
@@ -114,20 +86,5 @@ namespace MetX.Tests.Standard
             Assert.AreEqual("CoreCompile", modifier.Targets.AddSourceGeneratedFiles.AfterTargets);
             Assert.AreEqual(null, modifier.Targets.AddSourceGeneratedFiles.BeforeTargets);
         }
-
-        public static Modifier GetFullClient()
-        {
-            var filePath = FullClientXmlFilePath;
-            Assert.IsTrue(File.Exists(filePath));
-            return Modifier.LoadFile(filePath);
-        }
-
-        public static Modifier GetEmptyClient()
-        {
-            var filePath = EmptyClientXmlFilePath;
-            Assert.IsTrue(File.Exists(filePath));
-            return Modifier.LoadFile(filePath);
-        }
-
     }
 }
