+++ MetX/MetX.Tests/Standard/Generation/CSharp/Project/Pieces/Emit/EqualsTrue.xml
