+++ MetX/MetX.Tests/Standard/Generation/CSharp/Project/Pieces/FullClient.xml
