+++ MetX/MetX.Tests/Standard/Generation/CSharp/Project/Pieces/GenerateToPath/Missing.xml
