+++ MetX/MetX.Tests/Standard/Generation/CSharp/Project/Pieces/LangVersion/EqualsBlank.xml
