+++ MetX/MetX.Tests/Standard/Generation/CSharp/Project/Pieces/LangVersion/EqualsXyz.xml
